pyproject.toml
[project]
name = "docuquery"
version = "0.1.0"
description = ""
authors = [
    {name = "Jiehoon Lee",email = "jiehoonn@bu.edu"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "redis (>=7.1.0,<8.0.0)",
    "qdrant-client (>=1.16.2,<2.0.0)",
    "boto3 (>=1.42.33,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "openai (>=2.15.0,<3.0.0)",
    "anthropic (>=0.76.0,<0.77.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "google-genai (>=1.60.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
